Move books index spec tests to welcome page tests file
